pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "yahll"
version = "0.1.0"
description = "Self-evolving local AI coding agent"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "ddgs>=0.1",
    "pyperclip>=1.8.0",
    "fpdf2>=2.7.0",
    "Pillow>=10.0.0",
    "pyautogui>=0.9.54",
    "pygetwindow>=0.0.9",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "pystray>=0.19.0",
    "keyboard>=0.13.5",
    "faster-whisper>=1.0.0",
    "python-multipart>=0.0.9",
    "pywebview>=5.0.0",
    "mempalace @ git+https://github.com/milla-jovovich/mempalace.git",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[project.scripts]
yahll = "yahll.main:app"

[tool.pytest.ini_options]
env = ["PYTHONUTF8=1"]

[tool.setuptools.packages.find]
where = ["src"]
